video-editor/requirements.txt
moviepy>=2.0
numpy
opencv-python
pydub
flask
flask-cors
faster-whisper
mediapipe
rembg
pillow
pydantic>=2.0
anthropic>=0.40
